update facilitator page
